+++ examples/flat/src/imports.ts
@@ -1,0 +1,7 @@
+//import c from './exports';
+import { a, b } from './exports';
+import type { ScalarType, ObjType } from './exports';
+
+import path from 'path';
+import fs from 'node:fs';
+import console from 'console';